not sure how i missed the first one

+++ services/pic-sure-hpds/data/src/main/java/edu/harvard/hms/dbmi/avillach/hpds/data/genotype/FileBackedByteIndexedInfoStore.java
@@ -19,7 +19,7 @@ public class FileBackedByteIndexedInfoStore implements Serializable {
 	public final String column_key;
 	public final String description;
 	public boolean isContinuous;
-	public Double min = Double.MAX_VALUE, max = Double.MIN_VALUE;
+	public Float min = Float.MAX_VALUE, max = Float.MIN_VALUE;
 
 	public FileBackedByteIndexedStorage<String, String[]> allValues;
 	public TreeMap<Double, TreeSet<String>> continuousValueMap;
